refactor(widget-checkout): restructure to match widget conventions
Changes:
- Move mock-api/ out of src/ to package root
- Co-locate CheckoutDrawerContext with CheckoutDrawer at src/ root
- Dissolve onramp/ into providers/OnRampProvider/ with nested TransakProvider/ and MeshProvider/
- Move SelectSource components into pages/SelectSourcePage/
- Rename CheckoutWidgetContainer to AppProvider
