+++ packages/widget-checkout/src/index.ts
@@ -7,10 +7,10 @@ export {
   type CheckoutDrawerRef,
   LifiWidgetCheckout,
 } from './LifiWidgetCheckout.js'
-export type { TransakContextValue } from './onramp/transak/transakContext.js'
+export type { TransakContextValue } from './providers/OnRampProvider/TransakProvider/transakContext.js'
 export type {
   LoadedOnRampAdapter,
   OnRampProviderMeta,
-} from './onramp/types.js'
+} from './providers/OnRampProvider/types.js'
 export * from './types/checkout.js'
 export { checkoutNavigationRoutes } from './utils/navigationRoutes.js'
