+++ packages/widget-checkout/src/pages/SelectSourcePage/WalletConnectRow.tsx
